+++ apps/antalmanac/src/components/buttons/GitHub.tsx
@@ -1,0 +1,23 @@
+import { GitHub } from '@mui/icons-material';
+import { Button, Tooltip } from '@mui/material';
+
+import { GITHUB_LINK } from '$src/globals';
+
+const GitHubButton = () => {
+    return (
+        <Tooltip title="Check out our GitHub!">
+            <Button
+                color="inherit"
+                startIcon={<GitHub />}
+                size="large"
+                variant="text"
+                href={GITHUB_LINK}
+                target="_blank"
+            >
+                GitHub
+            </Button>
+        </Tooltip>
+    );
+};
+
+export default GitHubButton;
