+++ apps/antalmanac/src/components/buttons/Projects.tsx
@@ -1,0 +1,28 @@
+import { Button, SvgIcon, SvgIconProps, Tooltip } from '@mui/material';
+
+import { PROJECTS_LINK } from '$src/globals';
+
+export const ICSSCIcon = (props: SvgIconProps) => (
+    <SvgIcon viewBox="0 0 446 359" {...props}>
+        <path d="M205.278 7.60434C201.902 11.8259 199.37 18.299 199.37 22.2391C199.37 26.4606 191.211 42.2211 181.364 58.263C160.825 90.6282 131.566 147.478 124.813 168.023C122.563 175.622 118.061 193.071 115.248 206.862L110.184 231.909L106.807 213.897C104.838 204.047 100.337 189.412 96.679 181.532C86.2693 158.173 70.2327 145.508 74.1715 163.52C75.2969 168.586 74.7342 171.963 73.3275 171.963C67.982 171.963 63.4805 151.137 63.1991 127.778C62.9178 115.113 62.0737 103.856 60.667 102.73C57.2909 99.0713 33.9394 124.119 22.1229 143.82C-2.354 184.628 -7.13684 245.418 10.8692 287.915C20.9975 312.682 40.1289 337.167 57.8536 348.987C76.1409 361.089 95.2723 361.089 111.59 349.269C118.061 344.766 124.532 338.293 126.22 335.197C130.159 327.879 135.223 328.161 141.975 335.478C144.789 338.855 153.792 344.766 161.388 348.706C173.205 354.616 178.831 355.742 196.556 355.46C208.091 355.179 229.192 355.179 242.978 355.46C312.751 356.867 335.821 355.742 351.014 350.394C372.677 342.796 378.023 334.634 378.023 308.742C378.023 279.472 377.179 278.909 338.072 278.909C317.253 278.909 306.562 277.784 307.687 276.095C308.812 274.406 323.724 273.281 343.136 273.281C374.647 273.281 377.46 272.718 382.524 267.089C387.589 261.179 440.763 165.209 444.983 153.951C446.389 150.293 444.701 149.448 436.261 149.448H425.851L396.592 203.203L367.332 256.957L351.577 255.269C307.687 250.484 309.094 250.766 307.687 242.885C302.904 219.526 295.589 201.796 282.366 181.814C267.736 159.862 260.984 143.257 263.235 135.94C264.079 134.251 268.58 132.562 273.363 132.562C283.21 132.562 315.565 146.634 324.568 155.077C333.571 163.239 342.292 164.646 350.451 159.299C356.922 155.077 357.485 153.107 356.359 139.598C354.39 116.239 338.353 86.1252 314.72 61.3588C298.121 43.6283 290.244 37.9995 270.268 28.4307C256.764 21.9576 242.415 12.9517 238.195 8.44867C228.067 -2.52737 213.437 -2.80881 205.278 7.60434Z" />
+    </SvgIcon>
+);
+
+const ProjectsButton = () => {
+    return (
+        <Tooltip title="Learn more about our projects!">
+            <Button
+                color="inherit"
+                startIcon={<ICSSCIcon />}
+                size="large"
+                variant="text"
+                href={PROJECTS_LINK}
+                target="_blank"
+            >
+                ICSSC Projects
+            </Button>
+        </Tooltip>
+    );
+};
+
+export default ProjectsButton;
