+++ apps/antalmanac/src/globals.ts
@@ -2,5 +2,8 @@ export const DODGER_BLUE = '#1e90ff';
 export const BLUE = '#305db7';
 
 export const FEEDBACK_LINK = 'https://form.asana.com/?k=fZ3SGnuGknDmzTYdocgIUg&d=1208267282546207';
+export const DONATION_LINK = 'https://venmo.com/u/ICSSC';
+export const PROJECTS_LINK = 'https://studentcouncil.ics.uci.edu/projects';
+export const GITHUB_LINK = 'https://github.com/icssc/AntAlmanac';
 
 export const PLANNER_LINK = '/planner';
